pyproject.toml
[project]
name = "tadpoles-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
]
